extracted date time methods into a separate file
